feat: added reactivity to LyricsFile
- refactor: some module restructurization
