+++ chronograph/ui/sync_pages/wbw_sync_page.py
@@ -15,7 +15,7 @@ from chronograph.utils.file_backend.file_mutagen_mp4 import FileMP4
 from chronograph.utils.file_backend.file_mutagen_vorbis import FileVorbis
 from chronograph.utils.file_backend.file_untaggable import FileUntaggable
 from chronograph.utils.lyrics import Lyrics, LyricsFormat, LyricsHierarchyConversion
-from chronograph.utils.lyrics_file_helper import LyricsFile
+from chronograph.utils.lyrics import LyricsFile
 from chronograph.utils.player import Player
 from chronograph.utils.wbw.models.lyrics_model import LyricsModel
 from dgutils import Actions
